scripts/inference/tensorrt/requirements.txt
nvidia-tensorrt==8.4.0.6
pycuda
pyyaml
numpy
opencv-python-headless
